+++ server/service/src/invoice_line/outbound_shipment_unallocated_line/allocate/generate.rs
@@ -15,13 +15,13 @@ use crate::invoice_line::{
     outbound_shipment_unallocated_line::{
         DeleteOutboundShipmentUnallocatedLine, UpdateOutboundShipmentUnallocatedLine,
     },
-    stock_out_line::InsertInvoiceLine,
+    stock_out_line::InsertOutInvoiceLine,
 };
 
 #[derive(Default)]
 pub struct GenerateOutput {
     pub update_lines: Vec<UpdateOutboundShipmentLine>,
-    pub insert_lines: Vec<InsertInvoiceLine>,
+    pub insert_lines: Vec<InsertOutInvoiceLine>,
     pub update_unallocated_line: Option<UpdateOutboundShipmentUnallocatedLine>,
     pub delete_unallocated_line: Option<DeleteOutboundShipmentUnallocatedLine>,
     pub skipped_expired_stock_lines: Vec<StockLine>,
@@ -149,9 +149,9 @@ fn generate_new_line(
     invoice_id: &str,
     packs_to_allocate: f64,
     stock_line: &StockLine,
-) -> InsertInvoiceLine {
+) -> InsertOutInvoiceLine {
     let stock_line_row = &stock_line.stock_line_row;
-    InsertInvoiceLine {
+    InsertOutInvoiceLine {
         id: uuid::uuid(),
         invoice_id: invoice_id.to_string(),
         item_id: stock_line_row.item_id.clone(),
