+++ server/service/src/invoice_line/outbound_shipment_unallocated_line/allocate/mod.rs
@@ -4,8 +4,7 @@ use crate::{
             update_outbound_shipment_line, UpdateOutboundShipmentLine,
             UpdateOutboundShipmentLineError,
         },
-        stock_out_line::insert_stock_out_line,
-        stock_out_line::{InsertInvoiceLine, InsertInvoiceLineError},
+        stock_out_line::{insert_stock_out_line, InsertOutInvoiceLine, InsertOutInvoiceLineError},
         validate::check_line_exists_option,
     },
     service_provider::ServiceContext,
@@ -40,7 +39,7 @@ pub enum AllocateOutboundShipmentUnallocatedLineError {
     LineIsNotUnallocatedLine,
     // TODO NotThisStoreInvoice,
     // Internal
-    InsertOutboundShipmentLine(InputWithError<InsertInvoiceLine, InsertInvoiceLineError>),
+    InsertOutboundShipmentLine(InputWithError<InsertOutInvoiceLine, InsertOutInvoiceLineError>),
     UpdateOutboundShipmentLine(
         InputWithError<UpdateOutboundShipmentLine, UpdateOutboundShipmentLineError>,
     ),
